+++ packages/backend/jobs-manager/service/src/modules/notifications/emails/null-email.service.ts
@@ -1,0 +1,11 @@
+import { EmailRecipient, EmailService } from './email.service';
+import { ResetPasswordContext } from './templates/reset-password-context';
+
+export class MockedEmailService implements EmailService {
+  public sendResetPassword(
+    context: ResetPasswordContext,
+    recipients: EmailRecipient[],
+  ): Promise<void> {
+    return Promise.resolve();
+  }
+}
