+++ packages/backend/jobs-manager/service/src/modules/notifications/notifications.module.ts
@@ -1,11 +1,22 @@
 import { Module } from '@nestjs/common';
+import { JM_ENVIRONMENTS } from '../app.constants';
 import { EmailTemplatesProvider } from './emails/email-templates.provider';
-import { EmailService } from './emails/email.service';
+import { EmailService, MailJetEmailService } from './emails/email.service';
+import { MockedEmailService as NullEmailService } from './emails/null-email.service';
 
 @Module({
   imports: [],
   controllers: [],
-  providers: [EmailService, EmailTemplatesProvider],
+  providers: [
+    {
+      provide: EmailService,
+      useClass:
+        process.env.JM_ENVIRONMENT !== JM_ENVIRONMENTS.tests
+          ? MailJetEmailService
+          : NullEmailService,
+    },
+    EmailTemplatesProvider,
+  ],
   exports: [EmailService],
 })
 export class NotificationsModule {}
